refactor: 쇼핑 페이지 컴포넌트 폴더 수정
